+++ build_test/CHANGELOG.md
@@ -5,6 +5,9 @@
 - `resolveSources` and `testBuilder` now do a full `build_runner` build, with
   configuration as much as possible based on the some parameters.
 - Add `testBuilders` to run a test build with multiple builders.
+- Add `optionalBuilder` to `testBuilders` to have some builders be optional.
+- Add `testingBuilderConfig` to `testBuilders` to control builder config
+  override.
 - Add `resolvers` parameter to `testBuild` and `testBuilders`.
 - Breaking change: removed `tearDown` parameter to `resolveSources` for
   keeping resolvers across multiple tests.